Fixed Confidentiality and Availability

+++ dean/php/delete.php
@@ -1,13 +1,16 @@
 <?php
 $conn = mysqli_connect("localhost","root","","db_sis");
 
-$id = $_GET['id'];
-$sql = "delete from enroll where id_number = '$id'";
-$result = mysqli_query($conn,$sql);
-echo "<script>alert('Record Deleted');</script>";
-header("location:search.php");
-exit();
-
-
-
+session_start();
+if (!isset($_SESSION['username'])) {
+    header('Location: logout.php');
+    exit();
+} else {
+    $id = $_GET['id'];
+    $sql = "delete from enroll where id_number = '$id'";
+    $result = mysqli_query($conn,$sql);
+    echo "<script>alert('Record Deleted');</script>";
+    header("location:search.php");
+    exit();
+}
 ?>
